Set window.screen.width and window.screen.height
Hi @macbre !

I've seen a few websites using `window.screen.width` in JavaScript to detect a mobile, instead of using user-agent sniffing or something else. This isn't working when using Phantomas with the `--mobile` option.

`screen.width` is supposed to be the size of the physical screen, not the size of the browser's viewport. By default, PhantomJS is using the screen size of the computer on which it's running. I thing Phantomas should override this value and set it to the viewport size.

I implemented the following method: http://stackoverflow.com/questions/25045727/how-do-i-override-window-screen
Feel free to rearrange the code!